Finish splitting core
