+++ iris/__main__.py
@@ -2,23 +2,29 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this file,
 # You can obtain one at http://mozilla.org/MPL/2.0/.
 
-import coloredlogs
 import glob
+import logging
+import os
 import shutil
 import sys
 import tempfile
 from distutils.spawn import find_executable
 from multiprocessing import Process
 
+import coloredlogs
+import pytesseract
 
 import firefox.app as fa
 import firefox.downloader as fd
 import firefox.extractor as fe
-import test_runner
-from api.core import *
+from api.core_helper import get_module_dir, get_platform, get_run_id, get_image_debug_path
 from api.helpers.parse_args import parse_args
+from api.keys import Key
+from api.platform_iris import Platform
+from api.settings import Settings
 from firefox import cleanup
 from local_web import LocalWebServer
+from test_runner import run
 
 tmp_dir = None
 restore_terminal_encoding = None
@@ -32,6 +38,18 @@ coloredlogs.DEFAULT_FIELD_STYLES = {'levelname': {'color': 'cyan', 'bold': True}
 coloredlogs.DEFAULT_LEVEL_STYLES = {'warning': {'color': 'yellow', 'bold': True},
                                     'success': {'color': 'green', 'bold': True},
                                     'error': {'color': 'red', 'bold': True}}
+
+try:
+    os.stat(get_image_debug_path())
+except:
+    os.mkdir(get_image_debug_path())
+for debug_image_file in os.listdir(get_image_debug_path()):
+    file_path = os.path.join(get_image_debug_path(), debug_image_file)
+    try:
+        if os.path.isfile(file_path):
+            os.unlink(file_path)
+    except Exception as e:
+        continue
 
 
 def main(argv=None):
@@ -56,7 +74,7 @@ class Iris(object):
         self.start_local_web_server(self.local_web_root, self.args.port)
         self.main()
         self.create_run_directory()
-        test_runner.run(self)
+        run(self)
 
     def main(self, argv=None):
         global tmp_dir
@@ -322,6 +340,7 @@ class Iris(object):
             logger.critical('Cannot find required library 7zip, aborting Iris.')
             logger.critical('Please consult wiki for complete setup instructions.')
             self.finish(code=5)
+
 
 class RemoveTempDir(cleanup.CleanUp):
     """Class definition for cleanup helper responsible for deleting the temporary directory prior to exit."""
