+++ iris/api/settings.py
@@ -1,6 +1,9 @@
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this file,
+# You can obtain one at http://mozilla.org/MPL/2.0/.
 
-from helpers.parse_args import parse_args
 from core_helper import get_os, get_os_version
+from helpers.parse_args import parse_args
 from platform_iris import Platform
 
 DEFAULT_MIN_SIMILARITY = 0.8
